requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
appnope #@ file:///home/conda/feedstock_root/build_artifacts/appnope_1707233003401/work
asent==0.8.3
asttokens #@ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
backcall #@ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm #@ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
confection==0.1.4
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
cymem==2.0.8
datasets==2.10.1
debugpy #@ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_30hp2nowkm/croot/debugpy_1690905056188/work
decorator #@ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dill==0.3.6
executing #@ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastparquet==2024.2.0
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
huggingface-hub==0.9.1
idna==3.6
importlib_metadata #@ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
importlib_resources==6.4.0
ipykernel #@ file:///Users/runner/miniforge3/conda-bld/ipykernel_1708996616394/work
ipython #@ file:///Users/runner/miniforge3/conda-bld/ipython_1680185563420/work
ipywidgets==8.1.2
jedi #@ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client #@ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1710255804825/work
jupyter_core #@ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1669775234685/work
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lyricsgenius==3.0.1
marisa-trie==1.1.1
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline #@ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.14
murmurhash==1.0.10
nest_asyncio #@ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging #@ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pandas==2.2.1
parso #@ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
patsy==0.5.6
pexpect #@ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare #@ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
preshed==3.0.9
prompt-toolkit #@ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==5.26.1
psutil #@ file:///Users/runner/miniforge3/conda-bld/psutil_1705722557061/work
ptyprocess #@ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval #@ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==16.1.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments #@ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq #@ file:///Users/runner/miniforge3/conda-bld/pyzmq_1666828678061/work
ray==2.22.0
redis==5.0.4
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
responses==0.18.0
rpds-py==0.18.1
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
six #@ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spotipy==2.23.0
srsly==2.4.8
stack-data #@ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
statsmodels==0.14.2
sympy==1.12
tenacity==8.3.0
tensorboardX==2.6.2.2
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado #@ file:///Users/runner/miniforge3/conda-bld/tornado_1708363185581/work
tqdm==4.66.4
traitlets #@ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.21.2
tweetnlp==0.4.4
typer==0.9.4
typing_extensions #@ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata==2024.1
uritools==4.0.2
urlextract==1.9.0
urllib3==2.2.1
wasabi==1.1.2
wcwidth #@ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
weasel==0.3.4
widgetsnbextension==4.0.10
xxhash==3.4.1
yarl==1.9.4
zipp #@ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
